move /contract -> /wallet/contract
