+++ src/shared/app.tsx
@@ -25,9 +25,6 @@ import { ScrollToTop } from "./common/scroll-top";
 import { colors } from "./common/styles/style-color";
 import { fonts } from "./common/styles/style-font";
 import { TopBar } from "./common/top-bar";
-import { ChooseFunction } from "./contract/choose-function";
-import { Deploy } from "./contract/deploy";
-import { Interact } from "./contract/interact";
 import { Home } from "./home/home";
 import Wallet from "./wallet/wallet";
 
@@ -107,22 +104,6 @@ export class App extends Component<Props> {
               <Route exact path="/actions" component={Actions} />
               <Route exact path="/wallet" component={ComingSoon} />
               <Route path="/dev/wallet" component={Wallet} />
-
-              <Route
-                exact
-                path="/dev/wallet/smart-contract/deploy"
-                component={Deploy}
-              />
-              <Route
-                exact
-                path="/dev/wallet/smart-contract/interact"
-                component={Interact}
-              />
-              <Route
-                exact
-                path="/dev/wallet/smart-contract"
-                component={ChooseFunction}
-              />
               <Route component={NotFound} />
             </Switch>
           </ScrollToTop>
